1. enable conv mps with texture bias. 2. add scale mps. 3. remove duplicate computation in fetch kernel. 4. fix typo. (#1665)
Former-commit-id: bfca5725b5dc5667d26e23804f77c6d6e5cdf3e1

+++ metal/paddle-mobile-metallib/paddle-mobile-metallib/ConvAddReluMetal.metal
@@ -644,13 +644,13 @@ kernel void depthwise_conv_add_relu_3x3_half_winograd(texture2d_array<half, acce
     }
     
     if (param.hasAddOp == 1) {
-        half4 base = biasTexture.sample(sample, uint2(tx, ty), tc);
+        half4 base = biasTexture.sample(sample, float2(tx, ty), tc);
         res[0] += base;
-        base = biasTexture.sample(sample, uint2(tx + 1, ty), tc);
+        base = biasTexture.sample(sample, float2(tx + 1, ty), tc);
         res[1] += base;
-        base = biasTexture.sample(sample, uint2(tx, ty + 1), tc);
+        base = biasTexture.sample(sample, float2(tx, ty + 1), tc);
         res[2] += base;
-        base = biasTexture.sample(sample, uint2(tx + 1, ty + 1), tc);
+        base = biasTexture.sample(sample, float2(tx + 1, ty + 1), tc);
         res[3] += base;
     }
